feat(Generators): ✨ Place contact module behind feature Hacker module depends on contact. Contact module depends on internet Add contact feature to CI for testing Move contact data to contact module
Configure feature as part of alll.
