+++ src/data/mod.rs
@@ -4,7 +4,6 @@ pub mod beer;
 pub mod color;
 pub mod company;
 pub mod computer;
-pub mod contact;
 pub mod currency;
 pub mod datetime;
 pub mod files;
